pyproject.toml
[project]
name = "budget-app"
version = "0.1.0"
description = "Personal budget tracking API"
authors = [
    {name = "Mohammed Tarek Mehdi Gedidou",email = "mohammed.gedidou@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "sqlmodel (>=0.0.37,<0.0.38)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    # bcrypt 5+ on some environments breaks passlib 1.7.4 backend autodetect; pin to 3.x for now:
    "bcrypt (>=3.2.0,<4.0.0)"
]


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.25.0"
alembic = "^1.14.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
